Ej Clase 5
Agrego parte 2 de los ejercicios

+++ src/Ej4Clase/Producto.java
@@ -3,13 +3,14 @@ package Ej4Clase;
 public class Producto {
 	
 	private String nombre;
-	private int precio;
+	private double precio;
 	
-	public Producto(String nombre, int precio) {
+	public Producto(String nombre, double precio) {
 		
 		this.nombre=nombre;
 		this.precio=precio;
 	}
+
 
 	public String getNombre() {
 		return nombre;
@@ -19,11 +20,11 @@ public class Producto {
 		this.nombre = nombre;
 	}
 
-	public int getPrecio() {
+	public double getPrecio() {
 		return precio;
 	}
 
-	public void setPrecio(int precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
